CORE-7312: Persistence and Verification sandboxes should not have FLOW CPKs
Filter CPKs down to contract CPKs in VERIFICATION and PERSISTENCE type sandboxes.

Added an exception on empty sandbox with message
"No CPKs in this virtual node context. State and contract classes must be defined inside a contract CPK. ($vnc)".

Convert the following CPKs to contract CPKs:
- components/ledger/ledger-persistence/testing-datamodel
- testing/bundles/testing-cats
- testing/bundles/testing-dogs
- testing/bundles/testing-fish
- testing/cpbs/sandbox-security-manager-one
- testing/cpbs/sandbox-security-manager-two

Add these new contract CPKs:
- testing/cpbs/ledger-consensual-demo-contract
- testing/cpbs/ledger-utxo-demo-contract
