+++ settings.gradle
@@ -374,7 +374,9 @@ include 'testing:cpbs:crypto-custom-digest-one-consumer'
 include 'testing:cpbs:crypto-custom-digest-two-cpk'
 include 'testing:cpbs:crypto-custom-digest-two-consumer'
 include 'testing:cpbs:ledger-consensual-demo-app'
+include 'testing:cpbs:ledger-consensual-demo-contract'
 include 'testing:cpbs:ledger-utxo-demo-app'
+include 'testing:cpbs:ledger-utxo-demo-contract'
 include 'testing:cpbs:extendable-cpb'
 include 'testing:cpbs:test-cordapp'
 include 'testing:cpbs:test-cordapp-for-cache-testing'
