+++ testing/bundles/testing-dogs/build.gradle
@@ -12,7 +12,7 @@ group 'com.r3.testing'
 
 cordapp {
     targetPlatformVersion platformVersion.toInteger()
-    workflow {
+    contract {
         name 'Testing Dogs CPK'
         versionId 1
         vendor 'R3'
