+++ testing/cpbs/ledger-consensual-demo-contract/src/main/kotlin/net/cordapp/demo/consensual/contract/TestConsensualState.kt
@@ -1,0 +1,12 @@
+package net.cordapp.demo.consensual.contract
+
+import net.corda.v5.ledger.consensual.ConsensualState
+import net.corda.v5.ledger.consensual.transaction.ConsensualLedgerTransaction
+import java.security.PublicKey
+
+class TestConsensualState(
+    val testField: String,
+    override val participants: List<PublicKey>
+) : ConsensualState {
+    override fun verify(ledgerTransaction: ConsensualLedgerTransaction) {}
+}
