+++ testing/cpbs/ledger-utxo-demo-app/src/main/kotlin/net/cordapp/demo/utxo/FindTransactionFlow.kt
@@ -9,6 +9,7 @@ import net.corda.v5.base.util.contextLogger
 import net.corda.v5.crypto.SecureHash
 import net.corda.v5.ledger.utxo.UtxoLedgerService
 import net.corda.v5.ledger.utxo.transaction.UtxoLedgerTransaction
+import net.cordapp.demo.utxo.contract.TestUtxoState
 
 data class FindTransactionParameters(val transactionId: String)
 
@@ -21,7 +22,7 @@ data class UtxoTransactionResult(
 )
 
 // hacky DTO to enable JSON serialization without custom serializers
-fun UtxoDemoFlow.TestUtxoState.toResult(): TestUtxoStateResult {
+fun TestUtxoState.toResult(): TestUtxoStateResult {
     return TestUtxoStateResult(this.testField, this.participants.map { it.encoded })
 }
 
@@ -30,7 +31,7 @@ fun UtxoLedgerTransaction.toResult(): UtxoTransactionResult {
     return this.let {
         UtxoTransactionResult(
             it.id,
-            it.outputContractStates.map { (it as UtxoDemoFlow.TestUtxoState).toResult() },
+            it.outputContractStates.map { (it as TestUtxoState).toResult() },
             it.signatories.map { it.encoded })
     }
 }
