+++ testing/cpbs/ledger-utxo-demo-contract/src/main/kotlin/net/cordapp/demo/utxo/contract/TestContract.kt
@@ -1,0 +1,9 @@
+package net.cordapp.demo.utxo.contract
+
+import net.corda.v5.ledger.utxo.Contract
+import net.corda.v5.ledger.utxo.transaction.UtxoLedgerTransaction
+
+class TestContract : Contract {
+    override fun verify(transaction: UtxoLedgerTransaction) {
+    }
+}
