+++ testing/cpbs/ledger-utxo-demo-contract/src/main/kotlin/net/cordapp/demo/utxo/contract/TestUtxoState.kt
@@ -1,0 +1,11 @@
+package net.cordapp.demo.utxo.contract
+
+import net.corda.v5.ledger.utxo.BelongsToContract
+import net.corda.v5.ledger.utxo.ContractState
+import java.security.PublicKey
+
+@BelongsToContract(TestContract::class)
+class TestUtxoState(
+    val testField: String,
+    override val participants: List<PublicKey>
+) : ContractState
