+++ testing/persistence-testkit/src/main/kotlin/net/corda/db/persistence/testkit/helpers/Resources.kt
@@ -3,5 +3,5 @@ package net.corda.db.persistence.testkit.helpers
 object Resources {
     const val EXTENDABLE_CPB = "/META-INF/extendable-cpb.cpb"
     // adding this to test calculator CPB sandbox can't load entities defined in the above.
-    const val CALCULATOR_CPB = "/META-INF/calculator.cpb"
+    const val FISH_CPB = "/META-INF/testing-fish.cpb"
 }
